backend/requirements.txt
fastapi
uvicorn
python-dotenv
groq
chromadb
websockets
numpy
python-multipart

